requirements.txt
# MkDocs a Material téma
mkdocs>=1.6.0
mkdocs-material>=9.7.0

# Pluginy
mkdocs-glightbox>=0.4.0
mkdocs-encryptcontent-plugin>=3.0.0

# Závislosti pro glightbox
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Rozšíření pro Markdown (automaticky nainstalováno s Material)
pymdown-extensions>=10.2

# Cloudinary a správa prostředí
cloudinary>=1.44.0
python-dotenv>=1.2.0
